Исправлена отправка cors & csrf.

+++ src/main/java/ru/skypro/homework/repository/AdRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import ru.skypro.homework.entity.Ad;
 
 import java.util.List;
-public interface AdRepository extends JpaRepository<Ad, Integer> {
+public interface AdRepository extends JpaRepository<Ad, Long> {
     List<Ad> findAdsByUserId(Long id);
 }
